Iteration added. First case set to be solved
The iteration script is prepared, with only the renewable data missing (it is set to some dummy values just for the sake of implementation).

The data used is from the first day of the complete dataset, we can choose a more interesting day, or a group of them.

TO-DO: Renewable data, postprocessing of the data + expenditures, preparation of following cases, failure rate implementation...