Interrumpibility added
The function interrumpibility generates a simple report on the interrumpibility of service when there is a failure. It only creates the report line when the demand is not fully covered.

The cost column is cumulative, the final row has the total cost of the interruption of the system throughout the year.

To add: In order to check a more robust grid, we can add a counter of the failures that don't affect the demand coverage.